+++ setup/service.go
@@ -1,4 +1,4 @@
-package services
+package setup
 
 import (
 	"log"
